Unify redundant fixtures
