+++ tests/libs/conftest.py
@@ -1,5 +1,3 @@
 from raiden_contracts.tests.fixtures import *  # noqa
 from raiden_libs.test.fixtures.client import *  # noqa
 from raiden_libs.test.fixtures.web3 import *  # noqa
-
-from .fixture_overwrites import *  # noqa
