--- tests/libs/fixture_overwrites.py
@@ -1,8 +1,0 @@
-import pytest
-
-from raiden_contracts.contract_manager import ContractManager, contracts_precompiled_path
-
-
-@pytest.fixture(scope='session')
-def contracts_manager():
-    return ContractManager(contracts_precompiled_path())
