+++ tests/monitoring/fixtures/contracts.py
@@ -1,12 +1,5 @@
 import pytest
 
-from raiden_contracts.contract_manager import ContractManager, contracts_precompiled_path
-
-
-@pytest.fixture(scope='session')
-def contracts_manager():
-    return ContractManager(contracts_precompiled_path())
-
 
 @pytest.fixture
 def monitoring_service_contract(monitoring_service_external):
